+++ src/client/UnixStreamClientSockEP.cpp
@@ -45,11 +45,14 @@ UnixStreamClientSockEP::UnixStreamClientSockEP(std::string bindPath, std::string
 }
 
 // for server side client creation
-UnixStreamClientSockEP::UnixStreamClientSockEP() {}
+UnixStreamClientSockEP::UnixStreamClientSockEP() : ownsSocketFile_{false} {}
 
 UnixStreamClientSockEP::~UnixStreamClientSockEP()
 {
-	unlink(saddr_.sun_path);
+	if (ownsSocketFile_)
+	{
+		unlink(saddr_.sun_path);
+	}
 }
